// apps/web/src/client/api/model/showReelClip.ts
/**
 * Generated by orval v7.18.0 🍺
 * Do not edit manually.
 * ReeLuv API
 * ShowReelとVideoClipを管理するAPI
 * OpenAPI spec version: 1.0.0
 */

export interface ShowReelClip {
  id: string
  name: string
  /** @nullable */
  description: string | null
  videoStandard: string
  videoDefinition: string
  startTimecode: string
  endTimecode: string
  duration: string
}